+++ test/turbo_streamer/capture_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 class TurboStreamer::CaptureTest < ActiveSupport::TestCase
 
-  test "_capture" do
+  test "_capture returns the bytes it wrote, and writes them" do
     builder = TurboStreamer.new
     capture = nil
 
@@ -15,11 +15,8 @@ class TurboStreamer::CaptureTest < ActiveSupport::TestCase
       builder.key3 'value3'
     end
 
-    # Strip here because OJ adds new lines for unkown reasons
-    # TODO: this could break parsers that don't allowing trailing whitespace,
-    # it would be nice to configure or fix OJ to not output the newline
     assert_equal '"key2":"value2","key4":"value4"', capture.strip
-    assert_equal '{"key1":"value1","key3":"value3"}', builder.target!.strip
+    assert_equal '{"key1":"value1","key2":"value2","key4":"value4","key3":"value3"}', builder.target!.strip
   end
 
 end